feat: implement 2fa segment and and reorganize preferences structure
