+++ app/assets/javascripts/preferences/components/index.ts
@@ -1,0 +1,3 @@
+export * from './content';
+export * from './menu-item';
+export * from './pane';
