+++ app/assets/javascripts/preferences/index.tsx
@@ -1,0 +1,6 @@
+import { toDirective } from '../components/utils';
+import { PreferencesViewWrapper, PreferencesWrapperProps } from './view';
+
+export const PreferencesDirective = toDirective<PreferencesWrapperProps>(
+  PreferencesViewWrapper
+);
