+++ app/assets/javascripts/preferences/menu.tsx
@@ -1,17 +1,14 @@
+import { Icon, IconType } from '@/components/Icon';
 import { observer } from 'mobx-react-lite';
 import { FunctionComponent } from 'preact';
-import { PreferencesMenuItem } from '../PreferencesMenuItem';
-import { Preferences } from './preferences';
+import { MenuItem } from './components';
+import { Preferences } from './models/preferences';
 
-interface PreferencesMenuProps {
-  preferences: Preferences;
-}
-
-export const PreferencesMenu: FunctionComponent<PreferencesMenuProps> =
+export const PreferencesMenu: FunctionComponent<{ preferences: Preferences }> =
   observer(({ preferences }) => (
     <div className="min-w-55 overflow-y-auto flex flex-col px-3 py-6">
       {preferences.menuItems.map((pref) => (
-        <PreferencesMenuItem
+        <MenuItem
           key={pref.id}
           iconType={pref.icon}
           label={pref.label}
