+++ app/assets/javascripts/preferences/models/index.ts
@@ -1,0 +1,2 @@
+export * from './preferences';
+export * from './two-factor-auth';
