+++ app/assets/javascripts/preferences/panes/index.ts
@@ -1,0 +1,2 @@
+export * from './help-feedback';
+export * from './security';
